fix: implemented some more suggested refactoring
