+++ src/extension-host/extension-types/extension-manifest.model.ts
@@ -1,13 +1,5 @@
 import { ElevatedPrivilegeNames } from '@shared/models/elevated-privilege-names.model';
 
-/** Interface that stores the extension dependency information */
-export interface ExtensionDependency {
-  /** Extension id of the given extension dependency */
-  id: string;
-  /** Version of the given extension dependency */
-  version: string;
-}
-
 /** Information about an extension provided by the extension developer. */
 export type ExtensionManifest = {
   /** Name of the extension */
@@ -58,7 +50,7 @@ export type ExtensionManifest = {
    */
   activationEvents: string[];
   /** List of extension dependencies required for this extension to work */
-  extensionDependencies?: ExtensionDependency[];
+  extensionDependencies?: Record<string, string>;
   /** Path to the JSON file that defines the display data this extension is adding */
   displayData?: string;
   /** Id of publisher who published this extension on the extension marketplace */
