+++ src/shared/models/manage-extensions-privilege.model.ts
@@ -1,5 +1,4 @@
-import { ExtensionManifest } from '@extension-host/extension-types/extension-manifest.model';
-import { LanguageStrings } from 'platform-bible-utils';
+import { FullExtensionData } from './full-extension-data.model';
 
 /** Base64 encoded hash values */
 export type HashValues = Partial<{
@@ -12,38 +11,6 @@ export type ExtensionIdentifier = {
   extensionName: string;
   extensionVersion: string;
 };
-
-/** Interface that stores extension icon information */
-export interface ExtensionIcon {
-  /** Path to the icon's file. Could be a URL */
-  filepath: string;
-  /** Icon file extension (png, svg, ...) */
-  filetype: string;
-  /** Raw binary data of the icon, intended to be encoded in a base64 string */
-  data: string;
-  /** True if this icon was submitted as a URL, else false. */
-  isUrl: boolean;
-}
-
-/**
- * Full set of descriptive metadata for an extension including the extension manifest and
- * visualization data that is useful for the extension marketplace
- */
-export type FullExtensionData = Readonly<
-  Omit<ExtensionManifest, 'name' | 'version'> & {
-    id: string;
-    currentVersion: string;
-    displayName: LanguageStrings;
-    shortSummary: LanguageStrings;
-    description: LanguageStrings;
-    icon: ExtensionIcon;
-    locales: string[];
-    moreInfoUrl: string;
-    supportUrl: string;
-    fileSize: number;
-    hashcode: Record<string, string>;
-  }
->;
 
 /**
  * Represents all extensions that are installed. Note that packaged extensions cannot be disabled,
@@ -121,6 +88,9 @@ export type ManageExtensions = {
   disableExtension: DisableExtensionFunction;
   /** Function to retrieve details about all installed extensions */
   getInstalledExtensions: GetInstalledExtensionsFunction;
-  /** Function to retrieve full details about a list of installed or disabled extensions */
+  /**
+   * Function to retrieve descriptive metadata and visualization data for an extension. Useful for
+   * the extension marketplace.
+   */
   getExtensionsData: GetExtensionsDataFunction;
 };
